Completed

+++ database/migrations/2022_06_16_124954_create_cards_table.php
@@ -18,7 +18,7 @@ class CreateCardsTable extends Migration
             $table->string('title');
             $table->text('description', 500)->nullable();
             $table->string('thumb')->nullable();
-            $table->float('price')->nullable();
+            $table->decimal('price')->nullable();
             $table->string('series')->nullable();
             $table->string('sale_date')->nullable();
             $table->string('type')->nullable();
